🐍  Python Testing Setup (#110)
Currently, Python testing is handled as part of the wheel building
process. Every wheel is tested across with all available tests.
This significantly contributes to the time spent building wheels and is
much better left as an individual task.

This PR adds a proper Python testing setup to the CI of the project. To
this end, it adds `nox` sessions for determining coverage
```bash 
nox -rs coverage
```
and running the mypy linter
```bash
nox -rs mypy
```
Both of these steps, in combination with the regular tests (`nox -rs
tests`), are included in a new CI workflow.
This flow first checks the code base with the mypy linter and then
starts running the tests under all supported operating systems and a
suitable combination of Python versions. Furthermore, a separate job
determines the coverage and publishes it to codecov.

In the process of writing tests to increase the coverage, a small bug in
the parser for the calibration data from files was found and
subsequently fixed. Yay tests 🥳
Finally, all warnings reported by `mypy` have been addressed.

Signed-off-by: burgholzer <burgholzer@me.com>