👷 Streamline Z3 Configuration (#112)
QMAP's CI complexity mainly stems from some legacy best practises and
the huge burden of getting Z3 to work properly in a CI environment. This
is why I created https://github.com/cda-tum/setup-z3. An action that
properly sets up a version of Z3 by cloning the respective binary
release and providing the root directory as an environment variable.
As a result, a consistent version of Z3 can now be used throughout all
jobs and many jobs could be simplified again to just use a GitHub build
matrix.

This PR also significantly improves the capabilities of the
`FindZ3.cmake` script that can now find an appropriate Z3 version in
more cases than ever. It can even pick-up Z3 from the virtual
environment if it is installed from `pip`.
Furthermore, it now determines a proper version even when
cross-compiling or when multiple Z3 installations are present.

Last but not least, a wheel repair step is introduced for Windows wheels
which should increase the portability of the produced Python package.

Signed-off-by: burgholzer <burgholzer@me.com>