🐛🔧 Fix and Enhance Codecov Config (#154)
## Description

Up until today it has been a mystery why the coverage report for python
was reported as empty on codecov although it was not empty when
inspected. No more mystery... turns out codecov was configured to ignore
all files under `mqt/**/*`... which effectively turns all coverage
reports empty.

This PR fixes the underlying issue.
In addition, it uses the newly introduced *flag* feature of codecov to
split the coverage between the languages.
In the future, different thresholds could be applied to C++ and Python.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>