🚨🐛 Disable opinionated bugbear warning (#198)
## Description

`flake8-bugbear` recently introduced the following opinionated warning:

> B905: zip() without an explicit strict= parameter set. strict=True
causes the resulting iterator to raise a ValueError if the arguments are
exhausted at differing lengths. The strict= argument was added in Python
3.10, so don't enable this flag for code that should work on <3.10. For
more information: https://peps.python.org/pep-0618/

As stated, this only applies to Python 3.10+ and cannot be resolved for
older Python versions. As a consequence, we disable that check for the
foreseeable future.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>