🔥 Remove Z3 from build dependencies (#199)
## Description

This PR gets rid of the Z3 build dependency. Due to build isolation, the
virtual environment containing Z3 (that is being linked to the C++
binding library) does no longer exist after the build. Hence, the
package install is essentially broken if the Z3 Python package is picked
up during the build.

This was a workaround for getting the RtD build to work. I managed to
work around the underlying issue by figuring out a way to specify the
`Z3_ROOT` environment variable for the package installation step.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

Signed-off-by: Lukas Burgholzer <lukas.burgholzer@jku.at>