Relax ExactlyOne constraint for depth-optimal synthesis.
For depth-optimality it is sufficient to impose that at least one gate has to be present on each qubit. The transition
relation on the tableau itself constrains the gates enough already.

For the maxsat solution some adjustments have to be made because it relies on the exactlyone constraint to be present.