Restructuring of the heuristic system (#410)
## Description

This PR strives to solve multiple problems (that are just too
interconnected to keep them to separate PRs):
- 2 new heuristics `GateCountSumDistanceMinusSharedSwaps` and
`GateCountMaxDistanceOrSumDistanceMinusSharedSwaps`, where the former is
the sum of _all_ qubit pair distances minus an upper bound for how many
swaps could potentially be saved by sharing with other moving qubits,
and the latter is just the dominating heuristic over the previous
`GateCountMaxDistance` and the new
`GateCountSumDistanceMinusSharedSwaps`.
Tested on a subset of MQTbench and mapping to IBM Brisbane
`GateCountMaxDistanceOrSumDistanceMinusSharedSwaps` reduces the
effective branching rate by about 10% compared to
`GateCountMaxDistance`, while of course yielding the same results (if
lookahead is disabled) as both heuristics are principally admissible.
Unfortunately, the current default lookahead heuristic
`GateCountMaxDistance` seems to not work well with this new heuristic,
as the combination results in slightly higher costs. This is probably
due to scaling issues (since the new heuristic is closer to the real
cost and therefore larger, reducing the impact of the lookahead penalty)
but might be solvable with higher lookahead factors or a new better
suited lookahead heuristic.
- Introducing a more flexible system for heuristics (both for the main
heuristic and the lookahead heuristic). In the new system any
implementation specifics of a heuristic are isolated to a single
function calculating a heuristic value from a search node, outside of
those functions only a few characteristics of the heuristics are
relevant for the mapper:
- principal admissibility (i.e. admissibility at least on the optimal
solution path)
  - tightness (heuristics that are 0 in all goal nodes)
  - fidelity-awareness
- Fixing the handling of CNOT reversals in Dijkstra,
fixed-cost-calculation and all the pre-existing heuristics. The current
implementation resulted in both non-admissible heuristics and fixed
costs that did not accurately reflect the gates added by QMAP (the cost
calculation currently assumes that at most 1 reversal is added, while
QMAP actually inserts reversals for each backwards CNOT). This new
implementation is as robust as possible, only failing to produce
admissible heuristics in edge-cases where cumulative reversal costs on
one edge surpass SWAP costs resulting in a non-convex cost space (which
generally does not allow for heuristics, that are both admissible and
tight)
- Moving all methods from `HeuristicMapper::Node` to `HeuristicMapper`
(since they grew more and more dependent on data structures from
`HeuristicMapper` with recent PRs) and making them more atomic (i.e.
reducing points at which nodes are in an inconsistent state)
- Added support for semi-directional architectures (i.e. architectures
with both bidirectional and unidirectional edges)

List of minor changes and bug fixes:
- Cleaning up `HeuristicMapper::Node`:
  - Removing `nswaps` because of redundancy with `swaps.size()`
- Making `swaps` a one-dimensional vector (I assume the inner dimension
was originally intended for the possibility of adding multiple swaps per
node. However, this is currently not implemented, so all inner vectors
are just of length 1)
- renaming `done` to `validMapping`, which better describes the property
now that non-tight heuristics have been introduced to QMAP
- Fixed tracking of `HeuristicMapper::Node::validMappedTwoQubitGates`
and activated it also for non-fidelity-aware heuristics enabling a more
efficient check if a node has a valid mapping.
- Reduced the register sizes in the example circuits to the actual
number of qubits used in each circuit, to allow for easier checks in the
tests for the minimum required architecture size
- Made the ordering of validly mapped search nodes consistent (as
defined by `operator>`), where previously all validly mapped nodes with
the same total cost were considered equal (resulting in an arbitrary
ordering in the priority queue and thereby potentially different optimal
solutions for different principally admissible heuristics)
- fixing a bug in the handling of teleportation qubits in
`HeuristicMapper::createInitialMapping`, that causes this function to go
into an endless loop in some edge cases (instead of randomly iterating
over the whole coupling map, it only iterates over a part of it due to
the upper limit for the RNG being too low, which causes an endless loop
once that subgraph is fully mapped but free teleportation qubits remain)
- removing redundant `Mapper::fidelities`
- simplifying `Dijkstra::buildEdgeSkipTable` by internally calling
`Dijkstra::buildTable` for the 0th dimension
- moving checks for invalid settings from `HeuristicMapper::map` to new
method `HeuristicMapper::checkParameters`

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>