🚨 adapt for Qiskit 0.46.0 (#422)
## Description

This PR updates the project to work with the latest Qiskit version. In
turn, and in preparation for the 1.0 release of Qiskit, it adapts the
project to the expected removals already. While this raises the minimum
supported Qiskit version to 0.46.0, it should make switching to 1.0 as
easy as possible.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: burgholzer <burgholzer@me.com>