✨ New Zoned Neutral Atom Compiler (#624)
## Description

This rather big PR adds a new compiler for zoned
neutral atom architectures. The key benefit is its modular structure.
This should ease the further development of the compiler as only
components of the compiler must be touched.

This implementation in C++ is based on its previous implementation in
Python from [here](https://github.com/UCLA-VAST/ZAC) and its
corresponding article [here](https://arxiv.org/abs/2411.11784).

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are related to it.
- [x] I have added appropriate tests and documentation.
- [x] I have made sure that all CI jobs on GitHub pass.
- [x] The pull request introduces no new warnings and follows the
project's style guidelines.

---------

Signed-off-by: Yannick Stade <100073938+ystade@users.noreply.github.com>
Co-authored-by: pre-commit-ci[bot] <66853113+pre-commit-ci[bot]@users.noreply.github.com>
Co-authored-by: Lukas Burgholzer <burgholzer@me.com>