🐛 Deactivate PDF Export for Documentation and Prepare Patch Release (#659)
## Description

Curly braces in the Python code on the zoned neutral atom compiler page
of the documentation cause an issue in the PDF export via LaTeX on
readthedocs. Until we find another solution, this PR deactivates the PDF
export for the documentation.

Additionally, it corrects two typos on the page for the zoned neutral
atom compiler.

This PR also prepares the corresponding path release.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [ ] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes or removals.
- [x] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: Yannick Stade <100073938+ystade@users.noreply.github.com>