🐛 Fix CD (#661)
## Description

Since the addition of the Zoned Neutral Atom Compiler the CD keeps to
fail even though all tests succeed. The issue seems to be related to the
serialization of `chrono::micorseconds`. On many systems, this
serialization is not defined and was manually defined in
`include/na/zoned/Compiler.hpp`. Apparently, on the Linux systems used
for building the wheels, this serialization function is already defined.
To circumvent this issue, the type of the members in the statistics
structure was changed to `int64_t` returned by
`chrono::micorseconds::count()`, and the serialization was removed.

Additionally, this PR prepares the next patch release.

## Checklist:

<!---
This checklist serves as a reminder of a couple of things that ensure
your pull request will be merged swiftly.
-->

- [x] The pull request only contains commits that are focused and
relevant to this change.
- [ ] I have added appropriate tests that cover the new/changed
functionality.
- [x] I have updated the documentation to reflect these changes.
- [x] I have added entries to the changelog for any noteworthy
additions, changes, fixes or removals.
- [ ] I have added migration instructions to the upgrade guide (if
needed).
- [x] The changes follow the project's style guidelines and introduce no
new warnings.
- [x] The changes are fully tested and pass the CI checks.
- [x] I have reviewed my own code changes.

---------

Signed-off-by: Yannick Stade <100073938+ystade@users.noreply.github.com>